lvirt: improve to support clone host with multiple devices
Previously, we use "virt-copy-in" to copy customized configs into new created
VMs when doing VM setup or recovery, but "-d/domain" option may fail if
target guest vm has multiple disks (it's quit common on our CI clusters),
so here we add a post failhandle for this condition, try to copy the customized
configs to images directly if previous '-d/domain' failed.
And change the package name libguest-tools-c to libguest-tools, which is the
offical-suggest name, when install libguest related package (virt-copy-in here).

Change-Id: I5faebdfe64f98703d3ba38df0b6456ed0a64e43c
Signed-off-by: Gu Zheng <gzheng@ddn.com>