Improve the log monitoring on console
The console used to sleep 1 second each time after asking for log.
This patch changes the behavior. After got log, the console will
ask for the next one immediately. If the server has log, it will
reply immediately. If not, the sever wait for the next log.
Whenever a log is generated, the log will be sent to console
immediately.

Change-Id: I5a92b12b9451a3cd60c139605a0a73dc333aad4f
Signed-off-by: Li Xi <lixi@ddn.com>